Fixed up the map[string]EncapsulatedData to work with JSON
* This fix works for json Fixing, but points to a larger issue that I
  probably wnat to actually encode the data in a custom format. Now that
  we have more defined APIs this seems more reasonable

+++ integration-tests/limiter/overrides_test.go
@@ -38,7 +38,7 @@ func Test_Limiter_Overrides_Create(t *testing.T) {
 			Name:  "override1",
 			Limit: 32,
 			KeyValues: datatypes.KeyValues{
-				"other": datatypes.Float64(32),
+				"other": datatypes.Float32(32),
 			},
 		}
